dataset_generation/requirements.txt
warcio==1.7.1
newspaper3k==0.2.8
